+++ packages/web/src/pages/sign-in-page/SignInPageDesktop.tsx
@@ -7,18 +7,17 @@ import audiusLogoColored from 'assets/img/audiusLogoColored.png'
 import { HarmonyPasswordField } from 'components/form-fields/HarmonyPasswordField'
 import { HarmonyTextField } from 'components/form-fields/HarmonyTextField'
 import PreloadImage from 'components/preload-image/PreloadImage'
-import { useNavigateToPage } from 'hooks/useNavigateToPage'
 import {
   ArtworkContainer,
   AudiusValues
 } from 'pages/sign-on/components/AudiusValues'
 import { LeftContentContainer } from 'pages/sign-on/components/desktop/LeftContentContainer'
-import { MetaMaskOption } from 'pages/sign-on/components/desktop/MetaMaskOption'
 import { SignOnContainerDesktop } from 'pages/sign-on/components/desktop/SignOnContainerDesktop'
 import { userHasMetaMask } from 'pages/sign-up-page/utils/metamask'
-import { FEED_PAGE, SIGN_UP_PAGE } from 'utils/route'
+import { SIGN_UP_PAGE } from 'utils/route'
 
 import styles from './SignInPage.module.css'
+import { SignInWithMetaMaskButton } from './SignInWithMetaMaskButton'
 
 const messages = {
   title: 'Sign Into Audius',
@@ -30,18 +29,6 @@ const messages = {
 }
 
 export const SignInPageDesktop = () => {
-  const navigate = useNavigateToPage()
-
-  const handleSignInWithMetaMask = async () => {
-    try {
-      window.localStorage.setItem('useMetaMask', JSON.stringify(true))
-    } catch (err) {
-      console.error(err)
-    }
-    navigate(FEED_PAGE)
-    window.location.reload()
-  }
-
   return (
     <Flex h='100%' alignItems='center' justifyContent='center'>
       <SignOnContainerDesktop>
@@ -77,12 +64,7 @@ export const SignInPageDesktop = () => {
                       <Button iconRight={IconArrowRight} type='submit'>
                         {messages.signIn}
                       </Button>
-                      {userHasMetaMask ? (
-                        <MetaMaskOption
-                          text='Sign In With'
-                          onClick={handleSignInWithMetaMask}
-                        />
-                      ) : null}
+                      {userHasMetaMask ? <SignInWithMetaMaskButton /> : null}
                     </Flex>
                     <Flex direction='row' alignItems='flexStart'>
                       <Text color='heading'>
